Add files via upload
Created a graphical user interface for playing tic tac toe using java. Implement a graphical interface that allows two players to play. Added an "Artificial Intelligence" that plays it with at least 2 levels of difficulty. The program kept a log of all the previous matches and being able to load them if required.